mpi: move lowering to xds/transforms
