+++ xdsl/xdsl_opt_main.py
@@ -17,6 +17,7 @@ from xdsl.dialects.vector import Vector
 from xdsl.dialects.memref import MemRef
 from xdsl.dialects.llvm import LLVM
 from xdsl.dialects.irdl import IRDL
+from xdsl.transforms.mpi_to_llvm import mpi_to_llvm_lowering
 
 from xdsl.irdl_mlir_printer import IRDLPrinter
 from xdsl.utils.exceptions import DiagnosticException
@@ -211,7 +212,7 @@ class xDSLOptMain:
 
         Add other/additional passes by overloading this function.
         """
-        pass
+        self.available_passes['mpi-to-llvm'] = mpi_to_llvm_lowering
 
     def register_all_targets(self):
         """
